pyproject.toml
[project]
name = "todoapp"
version = "0.1.0"
description = "Multi-list task management app"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1",
    "markdown>=3.7",
    "bleach>=6.1",
    "gunicorn>=22.0",
    "django-ninja>=1.5.3",
]

[dependency-groups]
dev = [
    "playwright>=1.49",
    "pytest-django>=4.8",
    "pytest-playwright>=0.6",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "todoapp.settings"
pythonpath = ["."]
addopts = "--basetemp=codex_pytest_tmp/run"
markers = [
    "slow: marks tests as slow (toast timeout, drag-drop)",
]

[tool.black]
line-length = 88
